+++ engine/PIM-Assignment-Wizard/PIM-Assignment-Wizard.ps1
@@ -101,10 +101,10 @@ Write-Output "******************************************************************
 
     Write-host ""
     Write-host "[ 01 / $($MaxSteps) ] Building list of all Users in Entra ID ... Please Wait !"
-    $Global:Users_All_ID = Get-MgUser -all:$true
+    $Global:Users_All_ID = Get-PimAdminsFiltered
 
     Write-host "[ 02 / $($MaxSteps) ] Building list of all Groups in Entra ID ... Please Wait !"
-    $Global:Groups_All_ID = Get-MgGroup -all:$true
+    $Global:Groups_All_ID = Get-PimGroupsFiltered
 
     Write-host "[ 03 / $($MaxSteps) ] Building list of all PIM-Groups in Entra ID ... Please Wait !"
     $Global:PIM_Groups_Definitions_ID = $Global:Groups_All_ID | `